scripts

+++ contracts/CarolusNFTV1.sol
@@ -28,7 +28,7 @@ contract CarolusNFTV1 is
     // https://stackoverflow.com/questions/53926612/is-there-a-way-to-compress-a-string-into-a-smaller-string-with-reversibility
     mapping(uint256 => string) public contentMap;
     mapping(uint256 => address) public tokenToAuthorMap;
-    mapping(uint256 => uint) public tokenToTimestampMap;
+    mapping(uint256 => uint256) public tokenToTimestampMap;
 
     mapping(uint256 => uint256) public tokenToUpvotesMap;
     mapping(uint256 => uint256) public tokenToDownvotesMap;
@@ -78,8 +78,8 @@ contract CarolusNFTV1 is
         _safeMint(msg.sender, tokenId);
 
         contentMap[tokenId] = content;
-    tokenToAuthorMap[tokenId] = msg.sender;
-    tokenToTimestampMap[tokenId] =block.timestamp;
+        tokenToAuthorMap[tokenId] = msg.sender;
+        tokenToTimestampMap[tokenId] = block.timestamp;
     }
 
     // withdraw funds by moderator
